Refactor/zero copy parsing

+++ include/Order.hpp
@@ -25,6 +25,6 @@ struct Order {
   bool buy;
 
   Order() = delete;
-  Order(uint64_t p, uint32_t o, uint16_t s, Order_action a, bool b)
+  Order(uint64_t p, uint32_t o, uint32_t s, Order_action a, bool b)
       : price(p), order_id(o), size(s), action(a), buy(b) {}
 };
